requirements.txt
python-constraint~=1.4.0
Flask~=1.1.2
connexion~=2.7.0
connexion[swagger-ui]